fix: split the stream translator out and stop Codacy's unconfigured ESLint
The stream module was still over the 500 line file limit after the test
split, so the wire translators move to stream/translate.rs and the
module keeps the HTTP pump, with the public names re-exported so no
caller changes.

The remaining findings all came from Codacy's ESLint running without a
configuration, applying Flow and ES5 era rules that reject plain
TypeScript (type imports, ES modules, classes) and type-aware checks it
has no type information for. This repository lints with Biome and
typechecks with tsc in CI, so the unconfigured ESLint engine is
disabled in .codacy.yml rather than appeased.
