+++ .codacy.yml
@@ -1,3 +1,6 @@
+engines:
+  eslint:
+    enabled: false
 exclude_paths:
   - "e2e/**"
   - "Cargo.lock"
